// frontend/e-commerce-service-mocked/src/app/generated/model/userResponseDTO.ts
/**
 * e-commerce-service
 * API documentation for the e-commerce-service
 *
 * OpenAPI spec version: v1
 * Contact: placeholder@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface UserResponseDTO { 
    userID?: string;
    userName?: string;
    email?: string;
    role?: string;
}